requirements.txt
# Kafka & Data Processing
kafka-python==2.0.2
pandas==2.1.0
python-dotenv==1.0.0
tqdm==4.66.0
numpy==1.24.3

# Spark Streaming
pyspark==3.5.0

# Database
psycopg2-binary==2.9.9

# Machine Learning
scikit-learn==1.3.0

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Dashboard
streamlit==1.28.0
plotly==5.18.0